finally, add final version of minishell

+++ heredoc.c
@@ -6,7 +6,7 @@
 /*   By: vfrants <vfrants@student.42vienna.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 23:02:31 by vfrants           #+#    #+#             */
-/*   Updated: 2023/11/06 18:59:43 by vfrants          ###   ########.fr       */
+/*   Updated: 2023/11/06 19:55:12 by vfrants          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,29 @@
 #include <readline/readline.h>
 #include <stdio.h>
 #include <unistd.h>
+
+int	replace_dollar(char *iterate, char **result, t_data *data, int ex)
+{
+	const int	vlen = count_var_len(iterate);
+	char		*buffer;
+	char		*key;
+
+	if (vlen == 1 && ex)
+		return (*result = ft_strdup(""), vlen);
+	if (vlen == 1)
+		return (*result = ft_strdup("$"), vlen);
+	key = ft_substr(iterate, 1, vlen - 1);
+	if (key == NULL)
+		return (vlen);
+	buffer = get_variable_value(data, key);
+	free(key);
+	if (buffer == NULL)
+		*result = ft_strdup("");
+	else
+		*result = ft_strdup(buffer);
+	free(buffer);
+	return (vlen);
+}
 
 static char	*expand_local_token(t_data *data, char *value)
 {
@@ -44,39 +67,61 @@ static char	*expand_local_token(t_data *data, char *value)
 	return (res);
 }
 
-int	do_heredoc(t_data *data, t_cmd_table *cmd_table, int i)
+static char	*get_name_heredoc(int i)
 {
-	char		*stop;
-	static int	j = -1;
 	char		*itoa_res;
 	char		*name;
-	char		*line;
-	int			fd;
 
-	stop = cmd_table->redirs[i].value;
 	itoa_res = ft_itoa(i);
 	if (itoa_res == NULL)
-		return (MALLOC_F);
-	name = ft_strcat(".oouyhukuliyty", itoa_res);
+		return (NULL);
+	name = ft_strcat(".temp_heredoc", itoa_res);
 	free(itoa_res);
-	if (name == NULL)
-		return (MALLOC_F);
-	fd = open(name , O_CREAT | O_TRUNC | O_RDWR, 0666);
-	if (fd < 0)
-		return (print_error(data, "Minishell: open: ", 1));
-	while (++j >= 0)
+	return (name);
+}
+
+static int	do_while(t_data *data, int fd, char *stop)
+{
+	static int	j = 0;
+	char		*line;
+
+	while (1)
 	{
 		line = readline("> ");
 		if (line == NULL)
-			return (close(fd), ft_printf_fd(STDERR_FILENO, "minishell: warning: here-document at line %d delimited by end-of-file (wanted `%s')\n", j, stop));
+		{
+			ft_printf_fd(STDERR_FILENO, "minishell: warning: \
+here-document at line %d delimited by end-of-file \
+(wanted `%s')\n", j, stop);
+			break ;
+		}
 		if (ft_strcmp(line, stop) == 0)
-			break;
+			break ;
 		line = expand_local_token(data, line);
 		if (!line)
 			return (close(fd), MALLOC_F);
 		ft_putendl_fd(line, fd);
 		free(line);
+		j++;
 	}
+	return (SUCCESS);
+}
+
+int	do_heredoc(t_data *data, t_cmd_table *cmd_table, int i)
+{
+	char		*name;
+	int			status;
+	int			fd;
+
+	name = get_name_heredoc(i);
+	if (!name)
+		return (MALLOC_F);
+	fd = open(name, O_CREAT | O_TRUNC | O_RDWR, 0666);
+	if (fd < 0)
+		return (print_error(data, "Minishell: open: ", 1));
+	status = do_while(data, fd, cmd_table->redirs[i].value);
+	if (status == MALLOC_F)
+		return (MALLOC_F);
 	close(fd);
 	if (cmd_table->redirs[i].no_space != 3)
 		unlink(name);
